requirements.txt
pandas==2.0.1
numpy==1.23.1
sentence-transformers==2.2.0
faiss-cpu==1.7.3
transformers==4.24.0
spacy==3.5.0
sklearn==1.1.0
torch==1.12.0
lime==0.2.0.1
nltk==3.7